+++ packages/cli/src/commands/index.ts
@@ -15,6 +15,7 @@ import { deployCommand } from './deploy/command';
 import { devCommand } from './dev/command';
 import { dnsCommand } from './dns/command';
 import { domainsCommand } from './domains/command';
+import { edgeConfigCommand } from './edge-config/command';
 import { envCommand } from './env/command';
 import { flagsCommand } from './flags/command';
 import { gitCommand } from './git/command';
@@ -43,6 +44,7 @@ import { removeCommand } from './remove/command';
 import { routesCommand } from './routes/command';
 import { rollbackCommand } from './rollback/command';
 import { rollingReleaseCommand } from './rolling-release/command';
+import { sandboxCommand } from './sandbox/command';
 import { skillsCommand } from './skills/command';
 import { targetCommand } from './target/command';
 import { teamsCommand } from './teams/command';
@@ -75,6 +77,7 @@ const commandsStructs = [
   devCommand,
   dnsCommand,
   domainsCommand,
+  edgeConfigCommand,
   envCommand,
   flagsCommand,
   gitCommand,
@@ -101,6 +104,7 @@ const commandsStructs = [
   routesCommand,
   rollbackCommand,
   rollingReleaseCommand,
+  sandboxCommand,
   skillsCommand,
   targetCommand,
   teamsCommand,