+++ packages/cli/src/commands/teams/index.ts
@@ -2,12 +2,14 @@ import list from './list';
 import add from './add';
 import change from './switch';
 import invite from './invite';
+import request from './request';
 import members from './members';
 import { parseArguments } from '../../util/get-args';
 import {
   addSubcommand,
   inviteSubcommand,
   listSubcommand,
+  requestSubcommand,
   membersSubcommand,
   switchSubcommand,
   teamsCommand,
@@ -26,6 +28,7 @@ const COMMAND_CONFIG = {
   switch: ['switch', 'change'],
   add: ['create', 'add'],
   invite: ['invite'],
+  request: ['request', 'access-request'],
   members: ['members', 'member'],
 };
 
@@ -118,6 +121,15 @@ export default async function teams(client: Client) {
       telemetry.trackCliSubcommandInvite(subcommandOriginal);
       return invite(client, args);
     }
+    case 'request': {
+      if (needHelp) {
+        telemetry.trackCliFlagHelp('teams', subcommandOriginal);
+        printHelp(requestSubcommand);
+        return 2;
+      }
+      telemetry.trackCliSubcommandRequest(subcommandOriginal);
+      return request(client, args);
+    }
     case 'members': {
       if (needHelp) {
         telemetry.trackCliFlagHelp('teams', subcommandOriginal);
@@ -129,7 +141,7 @@ export default async function teams(client: Client) {
     }
     default: {
       output.error(
-        'Please specify a valid subcommand: add | ls | switch | invite | members'
+        'Please specify a valid subcommand: add | ls | switch | invite | request | members'
       );
       output.print(help(teamsCommand, { columns: client.stderr.columns }));
       return 2;