+++ packages/cli/src/util/telemetry/commands/project/index.ts
@@ -27,6 +27,20 @@ export class ProjectTelemetryClient
     });
   }
 
+  trackCliSubcommandAccessSummary(actual: string) {
+    this.trackCliSubcommand({
+      subcommand: 'access-summary',
+      value: actual,
+    });
+  }
+
+  trackCliSubcommandChecks(actual: string) {
+    this.trackCliSubcommand({
+      subcommand: 'checks',
+      value: actual,
+    });
+  }
+
   trackCliSubcommandRemove(actual: string) {
     this.trackCliSubcommand({
       subcommand: 'remove',
@@ -51,6 +65,27 @@ export class ProjectTelemetryClient
   trackCliSubcommandAccessGroups(actual: string) {
     this.trackCliSubcommand({
       subcommand: 'access-groups',
+      value: actual,
+    });
+  }
+
+  trackCliSubcommandProtection(actual: string) {
+    this.trackCliSubcommand({
+      subcommand: 'protection',
+      value: actual,
+    });
+  }
+
+  trackCliSubcommandWebAnalytics(actual: string) {
+    this.trackCliSubcommand({
+      subcommand: 'web-analytics',
+      value: actual,
+    });
+  }
+
+  trackCliSubcommandSpeedInsights(actual: string) {
+    this.trackCliSubcommand({
+      subcommand: 'speed-insights',
       value: actual,
     });
   }