+++ packages/express/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @vercel/express
 
+## 0.1.70
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/cervel@0.0.44
+  - @vercel/node@5.7.2
+
+## 0.1.69
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.7.1
+  - @vercel/cervel@0.0.43
+
 ## 0.1.68
 
 ### Patch Changes