+++ packages/hono/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vercel/hono
 
+## 0.2.63
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.7.2
+
+## 0.2.62
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.7.1
+
 ## 0.2.61
 
 ### Patch Changes