+++ packages/koa/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @vercel/koa
 
+## 0.1.43
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.7.2
+
+## 0.1.42
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @vercel/node@5.7.1
+
 ## 0.1.41
 
 ### Patch Changes